+++ src/editor.js
@@ -2,7 +2,7 @@ export default {
     "list.dropBackground": "#383b3d",
     "editorLineNumber.activeForeground": "#dddddd",
     "editor.background": "#1e1e1e",
-    "editor.foreground": "<text>",
+    "editor.foreground": "<white>",
     "editor.selectionHighlightBackground": "#add6ff26",
     "editor.inactiveSelectionBackground": "#add6ff26",
     "editorIndentGuide.activeBackground": "<gray>",
@@ -13,48 +13,48 @@ export default {
     "activityBarBadge.background": "#007acc",
     "sideBarTitle.foreground": "#dddddd",
     "editor.foldBackground": "#00000000",
-    "editorGutter.addedBackground": "<green>",
-    "editorOverviewRuler.addedForeground": "<green>", // Overview ruler marker color for added content.
+    "editorGutter.addedBackground": "<addition>",
+    "editorOverviewRuler.addedForeground": "<addition>", // Overview ruler marker color for added content.
     "editorGutter.modifiedBackground": "<escape>",
     "editorOverviewRuler.modifiedForeground": "<escape>", // Overview ruler marker color for modified content.
-    "editorGutter.deletedBackground": "<red>",
-    "editorOverviewRuler.deletedForeground": "<red>", // Overview ruler marker color for deleted content.
+    "editorGutter.deletedBackground": "<error>",
+    "editorOverviewRuler.deletedForeground": "<error>", // Overview ruler marker color for deleted content.
     // Error, Warning, & Info
     // Error
-    "editorError.foreground": "<red>", // Error squiggles
-    "editorOverviewRuler.errorForeground": "<red>", // Overview ruler marker color for errors.
+    "editorError.foreground": "<error>", // Error squiggles
+    "editorOverviewRuler.errorForeground": "<error>", // Overview ruler marker color for errors.
     // Warning
     "editorWarning.foreground": "<escape>", // Warning squiggles
     "editorOverviewRuler.warningForeground": "<escape>", // Overview ruler marker color for warnings.
     // Info
-    "editorInfo.foreground": "<local>", // Info squiggles
-    "editorOverviewRuler.infoForeground": "<local>", // Overview ruler marker color for infos.
+    "editorInfo.foreground": "<lightblue>", // Info squiggles
+    "editorOverviewRuler.infoForeground": "<lightblue>", // Overview ruler marker color for infos.
     // Git Colors
-    "gitDecoration.modifiedResourceForeground": "<green>",
-    "gitDecoration.addedResourceForeground": "<green>",
+    "gitDecoration.modifiedResourceForeground": "<addition>",
+    "gitDecoration.addedResourceForeground": "<addition>",
     "gitDecoration.deletedResourceForeground": "<escape>",
     "gitDecoration.untrackedResourceForeground": "<escape>",
     "gitDecoration.ignoredResourceForeground": "#808080",
-    "gitDecoration.conflictingResourceForeground": "<red>",
+    "gitDecoration.conflictingResourceForeground": "<error>",
     // Terminal Colors
     "terminal.selectionBackground": "#add6ff26",
     "terminalCursor.foreground": "#808080",
     "terminal.background": "#1e1e1e",
-    "terminal.foreground": "<text>",
+    "terminal.foreground": "<white>",
     "terminal.ansiBlack": "#808080",
     "terminal.ansiBrightBlack": "#808080",
-    "terminal.ansiRed": "<red>",
-    "terminal.ansiBrightRed": "<red>",
-    "terminal.ansiBlue": "<keyword>",
-    "terminal.ansiBrightBlue": "<local>",
-    "terminal.ansiGreen": "<green>",
+    "terminal.ansiRed": "<error>",
+    "terminal.ansiBrightRed": "<error>",
+    "terminal.ansiBlue": "<blue>",
+    "terminal.ansiBrightBlue": "<lightblue>",
+    "terminal.ansiGreen": "<addition>",
     "terminal.ansiBrightGreen": "<number>",
-    "terminal.ansiCyan": "<definition>",
-    "terminal.ansiBrightCyan": "<definition>",
+    "terminal.ansiCyan": "<addition>",
+    "terminal.ansiBrightCyan": "<addition>",
     "terminal.ansiYellow": "<escape>",
     "terminal.ansiBrightYellow": "<string>",
-    "terminal.ansiMagenta": "<controlflow>",
-    "terminal.ansiBrightMagenta": "<controlflow>",
-    "terminal.ansiWhite": "<text>",
-    "terminal.ansiBrightWhite": "<text>",
+    "terminal.ansiMagenta": "<purple>",
+    "terminal.ansiBrightMagenta": "<purple>",
+    "terminal.ansiWhite": "<white>",
+    "terminal.ansiBrightWhite": "<white>",
 };