+++ src/tokenColors/c.js
@@ -1,0 +1,39 @@
+export default {
+    "prefix": "source.c",
+    "suffix": "",
+    "<white>": [
+        "punctuation.definition.begin.bracket",
+        "punctuation.definition.end.bracket",
+    ],
+    "<gray>": [
+        "storage.modifier.array",
+        "string.quoted.other",
+        "string.quoted.double.include",
+        "meta.preprocessor.include",
+        "keyword.operator",
+        "punctuation.separator.pointer-access",
+    ],
+    "<gray>": [
+        "storage.modifier.array.bracket",
+    ],
+    "<blue>": [
+        "keyword.operator.sizeof",
+        "keyword.other.typedef",
+        "storage.type",
+        "constant.language",
+        "storage.modifier",
+    ],
+    "<number>": [
+        "keyword.other.unit.suffix.floating-point",
+        "keyword.other.unit.hexadecimal",
+    ],
+    "<yellow>": [
+        "entity.name.function",
+        "entity.name.function.preprocessor",
+        "entity.name.function.preprocessor",
+    ],
+    "<purple>": [
+        "keyword.control.directive",
+        "keyword.control",
+    ],
+}