+++ src/tokenColors/cs.js
@@ -1,0 +1,61 @@
+export default {
+    "prefix": "source.cs",
+    "suffix": "",
+    "<white>": [
+        "variable.other.alias",
+        "punctuation.curlybrace",
+        "punctuation.parenthesis",
+        "entity.name.variable",
+        "variable.other.readwrite",
+        "variable.other.object",
+        "punctuation.squarebracket",
+        "meta.interpolation",
+        "invalid",
+    ],
+    "italic": [
+        "invalid",
+    ],
+    "<gray>": [
+        "punctuation.accessor",
+        "punctuation.definition.typeparameters",
+        "punctuation.separator",
+        "punctuation.tilde",
+        "keyword.operator",
+        "",
+        "entity.name.type.namespace",
+        "punctuation.terminator",
+        "entity.name.label",
+        "comment.block.documentation string",
+    ],
+    "<blue>": [
+        "keyword.type",
+        "keyword.other",
+        "keyword.query",
+        "constant",
+        "storage.modifier",
+    ],
+    "<struct>": [
+        "entity.name.type.struct",
+    ],
+    "<enum>": [
+        "entity.name.type.enum",
+    ],
+    "<red>": [
+        "keyword.operator.arrow",
+    ],
+    "<green>": [
+        "entity.name.type.class",
+        "entity.name.type.alias",
+        "storage.type",
+        "entity.name.type.interface",
+        "entity.name.type.type-parameter",
+        "entity.name.type.delegate",
+    ],
+
+    "<yellow>": [
+        "entity.name.function",
+    ],
+    "<purple>": [
+        "keyword.control",
+    ],
+}