+++ src/tokenColors/html.js
@@ -1,22 +1,26 @@
 export default {
+    "prefix": "text.html",
+    "suffix": "",
+    "<blue>": [
+        "entity.name.tag",
+    ],
+    "<purple>": [
+        "entity.name.tag.script",
+        "meta.tag.metadata.script entity.name.tag",
+    ],
+    "<green>": [
+        "entity.name.tag.template",
+        "meta.tag.metadata.link entity.name.tag",
+    ],
+    "<red>": [
+        "entity.name.tag.style",
+        "meta.tag.metadata.style entity.name.tag",
+    ],
     "<gray>": [
-        "punctuation.definition.tag.begin.html",
-        "punctuation.definition.tag.html",
-        "punctuation.definition.tag.end.html",
+        "punctuation.separator.key-value",
+        "punctuation.definition.tag",
     ],
-    "<local>": [
-        "invalid.deprecated.entity.other.attribute-name.html",
-    ],
-    "<string>": [
-        "punctuation.definition.string.begin.html string.other.link.description.title.html",
-        "punctuation.definition.string.end.html string.other.link.description.title.html",
-        "string.other.link.description.title.html",
-        "punctuation.definition.string.begin.html",
-        "punctuation.definition.string.end.html",
-        "string.quoted.double.html",
-        "string.quoted.single.html",
-    ],
-    "<definition>": [
-        "meta.attribute.unrecognized.v-bind:href.html entity.other.attribute-name.html",
+    "<lightblue>": [
+        "entity.other.attribute-name",
     ],
 }