+++ tokenColors.js
@@ -1,7 +1,8 @@
-import clang from "./src/tokenColors/clang.js";
-import csharp from "./src/tokenColors/csharp.js";
+import c from "./src/tokenColors/c.js";
+import cs from "./src/tokenColors/cs.js";
 import css from "./src/tokenColors/css.js";
 import gen_typescript from "./src/tokenColors/gen_typescript.js";
+import gen_vue from "./src/tokenColors/gen_vue.js";
 import gen_xml from "./src/tokenColors/gen_xml.js";
 import go from "./src/tokenColors/go.js";
 import html from "./src/tokenColors/html.js";
@@ -9,15 +10,16 @@ import java from "./src/tokenColors/java.js";
 import javascript from "./src/tokenColors/javascript.js";
 import json from "./src/tokenColors/json.js";
 import markdown from "./src/tokenColors/markdown.js";
-import markup from "./src/tokenColors/markup.js";
+import python from "./src/tokenColors/python.js";
 import _common from "./src/tokenColors/_common.js";
 
 
 export default [
-	clang,
-	csharp,
+	c,
+	cs,
 	css,
 	gen_typescript,
+	gen_vue,
 	gen_xml,
 	go,
 	html,
@@ -25,6 +27,6 @@ export default [
 	javascript,
 	json,
 	markdown,
-	markup,
+	python,
 	_common,
 ]